requirements.txt
pandas
matplotlib
python-binance
pandas_ta
scikit-learn